requirements.txt
Flask>=3.0.0
qrcode[pil]>=7.4.2
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
Werkzeug>=3.0.0
